refactor: now split library components from `confidential-word-count` example
