--- confidentialstorm/common/src/main/java/ch/usi/inf/confidentialstorm/common/api/HistogramService.java
@@ -1,13 +1,0 @@
-package ch.usi.inf.confidentialstorm.common.api;
-
-import ch.usi.inf.confidentialstorm.common.api.model.HistogramSnapshotResponse;
-import ch.usi.inf.confidentialstorm.common.api.model.HistogramUpdateRequest;
-import ch.usi.inf.confidentialstorm.common.crypto.exception.EnclaveServiceException;
-import org.apache.teaclave.javasdk.common.annotations.EnclaveService;
-
-@EnclaveService
-public interface HistogramService {
-    void update(HistogramUpdateRequest update) throws EnclaveServiceException;
-
-    HistogramSnapshotResponse snapshot();
-}
