+++ confidentialstorm/common/src/main/java/ch/usi/inf/confidentialstorm/common/api/IServiceMessage.java
@@ -1,0 +1,6 @@
+package ch.usi.inf.confidentialstorm.common.api;
+
+import java.io.Serializable;
+
+public interface IServiceMessage extends Serializable {
+}
