--- confidentialstorm/common/src/main/java/ch/usi/inf/confidentialstorm/common/api/SplitSentenceService.java
@@ -1,11 +1,0 @@
-package ch.usi.inf.confidentialstorm.common.api;
-
-import ch.usi.inf.confidentialstorm.common.api.model.SplitSentenceRequest;
-import ch.usi.inf.confidentialstorm.common.api.model.SplitSentenceResponse;
-import ch.usi.inf.confidentialstorm.common.crypto.exception.EnclaveServiceException;
-import org.apache.teaclave.javasdk.common.annotations.EnclaveService;
-
-@EnclaveService
-public interface SplitSentenceService {
-    SplitSentenceResponse split(SplitSentenceRequest request) throws EnclaveServiceException;
-}
