--- confidentialstorm/common/src/main/java/ch/usi/inf/confidentialstorm/common/api/SpoutMapperService.java
@@ -1,11 +1,0 @@
-package ch.usi.inf.confidentialstorm.common.api;
-
-import ch.usi.inf.confidentialstorm.common.crypto.exception.EnclaveServiceException;
-import ch.usi.inf.confidentialstorm.common.crypto.model.EncryptedValue;
-import ch.usi.inf.confidentialstorm.common.topology.TopologySpecification;
-import org.apache.teaclave.javasdk.common.annotations.EnclaveService;
-
-@EnclaveService
-public interface SpoutMapperService {
-    EncryptedValue setupRoute(TopologySpecification.Component component, EncryptedValue entry) throws EnclaveServiceException;
-}
