--- confidentialstorm/common/src/main/java/ch/usi/inf/confidentialstorm/common/api/WordCountService.java
@@ -1,11 +1,0 @@
-package ch.usi.inf.confidentialstorm.common.api;
-
-import ch.usi.inf.confidentialstorm.common.api.model.WordCountRequest;
-import ch.usi.inf.confidentialstorm.common.api.model.WordCountResponse;
-import ch.usi.inf.confidentialstorm.common.crypto.exception.EnclaveServiceException;
-import org.apache.teaclave.javasdk.common.annotations.EnclaveService;
-
-@EnclaveService
-public interface WordCountService {
-    WordCountResponse count(WordCountRequest request) throws EnclaveServiceException;
-}
