--- confidentialstorm/common/src/main/java/ch/usi/inf/confidentialstorm/common/api/model/base/IServiceMessage.java
@@ -1,6 +1,0 @@
-package ch.usi.inf.confidentialstorm.common.api.model.base;
-
-import java.io.Serializable;
-
-public interface IServiceMessage extends Serializable {
-}
