+++ confidentialstorm/enclave/src/main/java/ch/usi/inf/confidentialstorm/enclave/service/bolts/ConfidentialBoltService.java
@@ -1,6 +1,6 @@
 package ch.usi.inf.confidentialstorm.enclave.service.bolts;
 
-import ch.usi.inf.confidentialstorm.common.api.model.base.IServiceMessage;
+import ch.usi.inf.confidentialstorm.common.api.IServiceMessage;
 import ch.usi.inf.confidentialstorm.common.crypto.exception.EnclaveServiceException;
 import ch.usi.inf.confidentialstorm.common.crypto.model.EncryptedValue;
 import ch.usi.inf.confidentialstorm.common.topology.TopologySpecification;
