+++ confidentialstorm/host/pom.xml
@@ -79,5 +79,11 @@
             <optional>true</optional>
             <scope>provided</scope>
         </dependency>
+        <dependency>
+            <groupId>ch.usi.inf.examples</groupId>
+            <artifactId>confidential-word-count-common</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 </project>
