+++ examples/confidential-word-count/common/pom.xml
@@ -5,7 +5,7 @@
 
     <parent>
         <groupId>ch.usi.inf.examples</groupId>
-        <artifactId>confidential-word-count</artifactId>
+        <artifactId>confidential_word_count</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
