+++ examples/confidential-word-count/common/src/main/java/ch/usi/inf/examples/confidential_word_count/common/api/SplitSentenceService.java
@@ -1,0 +1,11 @@
+package ch.usi.inf.examples.confidential_word_count.common.api;
+
+import ch.usi.inf.confidentialstorm.common.crypto.exception.EnclaveServiceException;
+import ch.usi.inf.examples.confidential_word_count.common.api.model.SplitSentenceRequest;
+import ch.usi.inf.examples.confidential_word_count.common.api.model.SplitSentenceResponse;
+import org.apache.teaclave.javasdk.common.annotations.EnclaveService;
+
+@EnclaveService
+public interface SplitSentenceService {
+    SplitSentenceResponse split(SplitSentenceRequest request) throws EnclaveServiceException;
+}
