+++ examples/confidential-word-count/enclave/pom.xml
@@ -98,7 +98,8 @@
                                 <configuration>
                                     <executable>bash</executable>
                                     <arguments>
-                                        <argument>${project.basedir}/../../../tools/javaenclave/build_app/make.sh</argument>
+                                        <argument>${project.basedir}/../../../tools/javaenclave/build_app/make.sh
+                                        </argument>
                                         <argument>${project.basedir}</argument>
                                         <argument>${org.apache.teaclave.javasdk.platform}</argument>
                                         <argument>${org.apache.teaclave.javasdk.enclave.privatePem.path}</argument>
@@ -182,6 +183,12 @@
             <version>1.0-rc6</version>
             <optional>true</optional>
             <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>ch.usi.inf.examples</groupId>
+            <artifactId>confidential-word-count-common</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <scope>compile</scope>
         </dependency>
     </dependencies>
 </project>
