+++ examples/confidential-word-count/host/src/main/resources/jokes.enc.json
